requirements.txt
streamlit==1.34.0
pandas==2.2.2
google-cloud-bigquery==3.23.0
python-dotenv==1.0.0
openai==1.3.0